Some cleanup while I was reading through the codebase (and learning Vulkan!)
1. Some member functions can be made `const`.
2. Avoid using C-style casting.
3. explicit constructors for constructor with 1 argument.
4. Use STL algorithms whenever possible. Also pre-allocate buffer before pushing.
5. Removed unreachable code (like `break` after `return` and `exit` after `return`).
6. Initialized many previously uninitialized variables.
7. Handle error return codes (mostly from freopen_s in Windows).
8. Avoid flushing with every print statement by replacing `std::endl` with "\n".
9. Some off-by-one error involving finding a random ratio algorithms. Mostly just avoid the warning by explicitly casting RAND_MAX to float.

Maybe some more...

+++ examples/inlineuniformblocks/inlineuniformblocks.cpp
@@ -30,7 +30,8 @@
 #define OBJ_DIM 0.025f
 
 float rnd() {
-	return ((float)rand() / (RAND_MAX));
+	return (static_cast<float>(rand()) / 
+			static_cast<float>(RAND_MAX));
 }
 
 class VulkanExample : public VulkanExampleBase
